Moved some classes to more appropriate packages.
