+++ server/src/main/java/consamables/auth/OAuthAuthenticator.java
@@ -6,7 +6,6 @@ import java.util.UUID;
 import org.joda.time.DateTime;
 import org.joda.time.Duration;
 
-import consamables.api.AccessToken;
 import consamables.api.User;
 import consamables.jdbi.AccessTokenDAO;
 import consamables.jdbi.UserDAO;
